Move general parts in qaoa into basic
