+++ qamomile/circuit/algorithm/__init__.py
@@ -2,6 +2,7 @@
 
 from .basic import (
     cz_entangling_layer,
+    superposition_vector,
     rx_layer,
     ry_layer,
     rz_layer,
@@ -21,20 +22,16 @@ from .qaoa import (
     hubo_qaoa_layers,
     hubo_qaoa_state,
     ising_cost,
-    phase_gadget,
     qaoa_layers,
     qaoa_state,
-    superposition_vector,
     x_mixer,
 )
 
 __all__ = [
     # QAOA
-    "phase_gadget",
     "ising_cost",
     "x_mixer",
     "qaoa_layers",
-    "superposition_vector",
     "qaoa_state",
     "hubo_ising_cost",
     "hubo_qaoa_layers",
@@ -44,6 +41,7 @@ __all__ = [
     "ry_layer",
     "rz_layer",
     "cz_entangling_layer",
+    "superposition_vector",
     # FQAOA
     "initial_occupations",
     "givens_rotation",
