+++ qamomile/circuit/algorithm/basic.py
@@ -10,6 +10,24 @@ import qamomile.circuit as qmc
 
 
 @qmc.qkernel
+def superposition_vector(
+    n: qmc.UInt,
+) -> qmc.Vector[qmc.Qubit]:
+    """Create a uniform superposition state by applying Hadamard to all qubits.
+
+    Args:
+        n (qmc.UInt): Number of qubits.
+
+    Returns:
+        qmc.Vector[qmc.Qubit]: Qubit register in the |+>^n state.
+    """
+    q = qmc.qubit_array(n, name="q")
+    for i in qmc.range(n):
+        q[i] = qmc.h(q[i])
+    return q
+
+
+@qmc.qkernel
 def rx_layer(
     q: qmc.Vector[qmc.Qubit],
     thetas: qmc.Vector[qmc.Float],
@@ -92,4 +110,46 @@ def cz_entangling_layer(
     n = q.shape[0]
     for i in qmc.range(n - 1):
         q[i], q[i + 1] = qmc.cz(q[i], q[i + 1])
+    return q
+
+
+@qmc.qkernel
+def phase_gadget(
+    q: qmc.Vector[qmc.Qubit],
+    indices: qmc.Vector[qmc.UInt],
+    angle: qmc.Float,
+) -> qmc.Vector[qmc.Qubit]:
+    """Apply exp(-i * angle/2 * Z_{i0} Z_{i1} ... Z_{ik-1}).
+
+    Decomposes a k-body Z-rotation into CX + RZ primitives.
+
+    Args:
+        q (qmc.Vector[qmc.Qubit]): Qubit register.
+        indices (qmc.Vector[qmc.UInt]): Qubit indices for the interaction term.
+            Must be non-empty.
+        angle (qmc.Float): Rotation angle in radians.
+
+    Returns:
+        qmc.Vector[qmc.Qubit]: Updated qubit register.
+    """
+    # Preconditions: indices must be non-empty.
+    k = indices.shape[0]
+    last = k - 1
+    # CX forward ladder
+    for step in qmc.range(last):
+        next_step = step + 1
+        left = indices[step]
+        right = indices[next_step]
+        q[left], q[right] = qmc.cx(q[left], q[right])
+
+    # Apply RZ on the last qubit
+    q[indices[last]] = qmc.rz(q[indices[last]], angle=angle)
+
+    # CX reverse ladder (using positive-step range for qkernel compatibility)
+    for step in qmc.range(last):
+        rev = k - step - 2
+        rev_next = rev + 1
+        left = indices[rev]
+        right = indices[rev_next]
+        q[left], q[right] = qmc.cx(q[left], q[right])
     return q
