+++ qamomile/circuit/algorithm/qaoa.py
@@ -2,23 +2,7 @@ from __future__ import annotations
 
 import qamomile.circuit as qmc
 
-
-@qmc.qkernel
-def superposition_vector(
-    n: qmc.UInt,
-) -> qmc.Vector[qmc.Qubit]:
-    """Create a uniform superposition state by applying Hadamard to all qubits.
-
-    Args:
-        n (qmc.UInt): Number of qubits.
-
-    Returns:
-        qmc.Vector[qmc.Qubit]: Qubit register in the |+>^n state.
-    """
-    q = qmc.qubit_array(n, name="q")
-    for i in qmc.range(n):
-        q[i] = qmc.h(q[i])
-    return q
+from . import basic as _basic
 
 
 @qmc.qkernel
@@ -129,50 +113,8 @@ def qaoa_state(
     Returns:
         qmc.Vector[qmc.Qubit]: QAOA state vector.
     """
-    q = superposition_vector(n)
+    q = _basic.superposition_vector(n)
     q = qaoa_layers(p, quad, linear, q, gammas, betas)
-    return q
-
-
-@qmc.qkernel
-def phase_gadget(
-    q: qmc.Vector[qmc.Qubit],
-    indices: qmc.Vector[qmc.UInt],
-    angle: qmc.Float,
-) -> qmc.Vector[qmc.Qubit]:
-    """Apply exp(-i * angle/2 * Z_{i0} Z_{i1} ... Z_{ik-1}).
-
-    Decomposes a k-body Z-rotation into CX + RZ primitives.
-
-    Args:
-        q (qmc.Vector[qmc.Qubit]): Qubit register.
-        indices (qmc.Vector[qmc.UInt]): Qubit indices for the interaction term.
-            Must be non-empty.
-        angle (qmc.Float): Rotation angle in radians.
-
-    Returns:
-        qmc.Vector[qmc.Qubit]: Updated qubit register.
-    """
-    # Preconditions: indices must be non-empty.
-    k = indices.shape[0]
-    last = k - 1
-    # CX forward ladder
-    for step in qmc.range(last):
-        next_step = step + 1
-        left = indices[step]
-        right = indices[next_step]
-        q[left], q[right] = qmc.cx(q[left], q[right])
-
-    # Apply RZ on the last qubit
-    q[indices[last]] = qmc.rz(q[indices[last]], angle=angle)
-
-    # CX reverse ladder (using positive-step range for qkernel compatibility)
-    for step in qmc.range(last):
-        rev = k - step - 2
-        rev_next = rev + 1
-        left = indices[rev]
-        right = indices[rev_next]
-        q[left], q[right] = qmc.cx(q[left], q[right])
     return q
 
 
@@ -204,7 +146,7 @@ def hubo_ising_cost(
     """
     q = ising_cost(quad, linear, q, gamma)
     for indices, coeff in qmc.items(higher):
-        q = phase_gadget(q, indices, coeff * gamma)
+        q = _basic.phase_gadget(q, indices, coeff * gamma)
     return q
 
 
@@ -274,6 +216,6 @@ def hubo_qaoa_state(
     Returns:
         qmc.Vector[qmc.Qubit]: HUBO QAOA state vector.
     """
-    q = superposition_vector(n)
+    q = _basic.superposition_vector(n)
     q = hubo_qaoa_layers(p_val, quad, linear, higher, q, gammas, betas)
     return q
